requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
selenium==4.18.1
psutil==5.9.8
pydantic==2.6.1
python-multipart==0.0.9
requests==2.31.0
webdriver-manager==4.0.1
python-dotenv==1.0.1
openpyxl==3.1.2
pandas==2.0.3
beautifulsoup4==4.12.3
gspread==6.0.2
google-auth==2.28.1
google-cloud-bigquery==3.17.1
google-cloud-logging==3.10.0
